refactor: BugSearchRepository 위치 변경
