+++ src/main/java/com/moabam/api/application/bug/BugService.java
@@ -11,8 +11,8 @@ import org.springframework.transaction.annotation.Transactional;
 import com.moabam.api.application.member.MemberService;
 import com.moabam.api.domain.bug.BugHistory;
 import com.moabam.api.domain.bug.BugType;
+import com.moabam.api.domain.bug.repository.BugHistorySearchRepository;
 import com.moabam.api.domain.member.Member;
-import com.moabam.api.domain.repository.BugHistorySearchRepository;
 import com.moabam.api.dto.TodayBugResponse;
 import com.moabam.api.dto.bug.BugResponse;
 import com.moabam.global.common.util.ClockHolder;
